requirements.txt
easydict==1.10
einops==0.6.0
jsonlines==3.1.0
PyYAML==6.0
tensorboardX==2.5.1
tqdm==4.64.1
transformers==4.24.0
